fix: split into different catalog base url (#335)
* fix: split into different catalog base url

* fix: reduce to lowercase catalog
