+++ cumulus_lambda_functions/uds_api/collections_api.py
@@ -273,52 +273,3 @@ async def query_collections(request: Request, collection_id: Union[str, None] = 
     if collections_result['statusCode'] == 200:
         return collections_result['body']
     raise HTTPException(status_code=collections_result['statusCode'], detail=collections_result['body'])
-
-@router.get("/catalog")
-@router.get("/catalog/")
-async def get_catalog(request: Request, limit: Union[int, None] = 10, offset: Union[int, None] = 0, ):
-    LOGGER.debug(f'starting query_collections request: {request}')
-
-    authorizer: UDSAuthorizorAbstract = UDSAuthorizerFactory() \
-        .get_instance(UDSAuthorizerFactory.cognito,
-                      es_url=os.getenv('ES_URL'),
-                      es_port=int(os.getenv('ES_PORT', '443'))
-                      )
-    auth_info = FastApiUtils.get_authorization_info(request)
-    uds_collections = UdsCollections(es_url=os.getenv('ES_URL'),
-                                     es_port=int(os.getenv('ES_PORT', '443')))
-    collection_regexes = authorizer.get_authorized_collections(DBConstants.read, auth_info['ldap_groups'])
-    LOGGER.info(f'collection_regexes: {collection_regexes}')
-    authorized_collections = uds_collections.get_collections(collection_regexes)
-    LOGGER.info(f'authorized_collections: {authorized_collections}')
-    # Example: [{'collection_id': 'URN:NASA:UNITY:MAIN_PROJECT:DEV:CUMULUS_DAPA_UNIT_TEST___1697606545', 'granule_count': 0, 'start_time': 1697581345446, 'end_time': 1697581345446}]
-    collection_id = [k[DBConstants.collection_id] for k in authorized_collections]
-    LOGGER.info(f'authorized_collection_ids: {collection_id}')
-    # NOTE: 2022-11-21: only pass collections. not versions
-
-    try:
-        custom_params = {}
-        if limit > CollectionDapaQuery.max_limit:
-            LOGGER.debug(f'incoming limit > {CollectionDapaQuery.max_limit}. resetting to max. incoming limit: {limit}')
-            limit = CollectionDapaQuery.max_limit
-            custom_params['limit'] = limit
-        LOGGER.debug(f'new limit: {limit}')
-        pg_link_generator = PaginationLinksGenerator(request, custom_params)
-
-        catalog = Catalog(
-            id='unity_ds',
-            description='Unity DS Catalog',
-            title='Unity DS Catalog',
-            href=pg_link_generator.base_url
-        )
-        authorized_collections_links = [Link(
-            rel='child',
-            target=f'{pg_link_generator.base_url}/collections/{k["collection_id"]}',
-            media_type='application/json',
-            title=k["collection_id"],
-        ) for k in authorized_collections]
-        catalog.add_links(authorized_collections_links)
-        return catalog.to_dict(True, False)
-    except Exception as e:
-        LOGGER.exception('failed during get_granules_dapa')
-        raise HTTPException(status_code=500, detail=str(e))
