+++ cumulus_lambda_functions/uds_api/web_service_constants.py
@@ -1,6 +1,7 @@
 class WebServiceConstants:
     API_PREFIX = 'am-uds-dapa'
     COLLECTIONS = 'collections'
+    CATALOG = 'catalog'
     DEPLOYED_STAGE = 'deployed_stage'
     BASE_URL = 'UDS_BASE_URL'
     ITEMS = 'items'
